refactor(solver): own narrowing cache guard state (#14347)
