+++ crates/tsz-solver/src/narrowing/compound.rs
@@ -72,8 +72,8 @@ impl<'a> NarrowingContext<'a> {
         // re-mints `T & narrowed` and recurses here, so a self-referential
         // constraint must be bounded on this path too. Charge one unit per call
         // and bail to the unchanged source when the budget is spent.
-        let _frame = self.enter_exclusion_frame();
-        if !self.charge_exclusion_work() {
+        let _frame = self.cache.enter_exclusion_frame();
+        if !self.cache.charge_exclusion_work() {
             return source_type;
         }
